requirements.txt
# Core Streamlit and UI
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.6

# AI and Language Processing
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.85.0

# Document Processing
jinja2>=3.1.0
python-docx>=1.1.0
reportlab>=4.0.0
Pillow>=10.0.0

# HTTP and Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.1.0

# Configuration and Utilities
python-dotenv>=1.0.0
pydantic>=2.9.0
typing-extensions>=4.11.0

# Additional UI Enhancements
streamlit-ace>=0.1.1
streamlit-aggrid>=0.3.4
streamlit-authenticator>=0.2.3
